requirements.txt
pystray>=0.19.0
Pillow>=9.0
certifi
truststore
